MV: entry routes and cohort routes to own file
