+++ routes/cohortRoutes.js
@@ -1,0 +1,21 @@
+var routes = {};
+var path = require('path');
+var Cohort = require(path.join(__dirname, '../models/cohort'));
+
+routes.POSTnewCohortEntry = function(req, res, next) {
+  var cohortName = req.body.name;
+  var cohortComment = req.body.comment;
+
+  Cohort.create({
+    name: cohortName,
+    comment: cohortComment
+  }, function(err, newCohortEntryObj) {
+    if (err) {
+      res.send(err)
+    }
+    console.log("new cohort entry", newCohortEntryObj)
+    res.json(newCohortEntryObj)
+  })
+};
+
+module.exports = routes;
